Enhancement: Extract IndexAction from TalksController
