+++ classes/Http/Controller/Admin/TalksController.php
@@ -13,9 +13,6 @@ declare(strict_types=1);
 
 namespace OpenCFP\Http\Controller\Admin;
 
-use OpenCFP\Domain\Services\Authentication;
-use OpenCFP\Domain\Services\Pagination;
-use OpenCFP\Domain\Talk\TalkFilter;
 use OpenCFP\Domain\Talk\TalkHandler;
 use OpenCFP\Domain\ValidationException;
 use OpenCFP\Http\Controller\BaseController;
@@ -28,66 +25,15 @@ use Twig_Environment;
 class TalksController extends BaseController
 {
     /**
-     * @var Authentication
-     */
-    private $authentication;
-
-    /**
-     * @var TalkFilter
-     */
-    private $talkFilter;
-
-    /**
      * @var TalkHandler
      */
     private $talkHandler;
 
-    public function __construct(
-        Authentication $authentication,
-        TalkFilter $talkFilter,
-        TalkHandler $talkHandler,
-        Twig_Environment $twig,
-        UrlGeneratorInterface $urlGenerator
-    ) {
-        $this->authentication = $authentication;
-        $this->talkFilter     = $talkFilter;
-        $this->talkHandler    = $talkHandler;
+    public function __construct(TalkHandler $talkHandler, Twig_Environment $twig, UrlGeneratorInterface $urlGenerator)
+    {
+        $this->talkHandler = $talkHandler;
 
         parent::__construct($twig, $urlGenerator);
-    }
-
-    public function indexAction(Request $request): Response
-    {
-        $adminUserId = $this->authentication->user()->getId();
-        $options     = [
-            'order_by' => $request->get('order_by'),
-            'sort'     => $request->get('sort'),
-        ];
-
-        $formattedTalks = $this->talkFilter->getTalks(
-            $adminUserId,
-            $request->get('filter'),
-            $options
-        );
-
-        // Set up our page stuff
-        $perPage    = (int) $request->get('per_page') ?: 20;
-        $pagerfanta = new Pagination($formattedTalks, $perPage);
-
-        $pagerfanta->setCurrentPage($request->get('page'));
-        $pagination = $pagerfanta->createView('/admin/talks?', $request->query->all());
-
-        return $this->render('admin/talks/index.twig', [
-            'pagination'   => $pagination,
-            'talks'        => $pagerfanta->getFanta(),
-            'page'         => $pagerfanta->getCurrentPage(),
-            'current_page' => $request->getRequestUri(),
-            'totalRecords' => \count($formattedTalks),
-            'filter'       => $request->get('filter'),
-            'per_page'     => $perPage,
-            'sort'         => $request->get('sort'),
-            'order_by'     => $request->get('order_by'),
-        ]);
     }
 
     public function viewAction(Request $request): Response
